Names Changes and kwarg handling has changed
read, write and group kwargs passed to xlmp and xlsmp now must be separate dicts.  README IS NOW OUT OF DATE.  Also Ixl renamed to _ExcelInterface.  An interface hierarchy will be needed for interfaces to something other than xlrd and xlwt.  This is github commit so test suite has not been ran.